Create an icon

Create an icon for usage in createRating.

createIcon(
  id,
  number = 10,
  rating_id,
  cumul = FALSE,
  tlp = FALSE,
  read_only = FALSE,
  hover = FALSE,
  i_name = "star",
  i_lib = "fontawesome",
  i_number = 1,
  i_on_color = "primary",
  i_off_color = "dark",
  init_color = "dark",
  i_width = NULL,
  i_height = NULL,
  i_anim = "none",
  i_duration = 2,
  i_value = 1,
  i_margin_left = "auto",
  i_margin_right = "auto",
  tlp_position = "top",
  tlp_color = "black"
)

Arguments

id

The icon id.

number

The total number of icons in the rating widget.

rating_id

The rating widget id.

cumul

Logical, whether or not to use cumulative color for the icon.

tlp

Logical, whether or not to use a tooltip for the icon.

read_only

Logical, whether or not to use a read only mode for the icon.

hover

Logical, whether or not to use a hover mode for the icon.

i_name

The icon name (fontawesome) or path (local).

i_lib

The icon library to use. Either "fontawesome" or "local".

i_number

The icon place in the rating widget.

i_on_color

The icon color when selected (hex or Bootstrap 5 color).

i_off_color

The icon color when deselected (hex or Bootstrap 5 color).

init_color

The icon initial color (hex or Bootstrap 5 color).

i_width

The icon width in css units.

i_height

The icon height in css units.

i_anim

The type of animation see Animate.css.

i_duration

The animation duration in seconds.

i_value

The character string to be shown as message.

i_margin_left, i_margin_right

The length value for the margin that's either left or right of the icon. By default, "auto" is used for both properties. If space is needed on either side then a length of "0.2em" is recommended as a starting point.

tlp_position

The tooltip position, "top", "right", "bottom", or "left".

tlp_color

The tooltip color, "primary", "secondary", "light", "dark", "info", "warning", "danger", "success", "white", or "black".

Value

A shiny.tag object for usage in createRating