Create rating icons

Create rating icons for usage in ratingInput.

Usage

createRating(
  rating_id,
  i_name = "star",
  i_lib = "fontawesome",
  number = 10,
  init_value = 0,
  i_on_color = "primary",
  i_off_color = "dark",
  i_width = NULL,
  i_height = NULL,
  i_anim = "none",
  i_duration = 2,
  cumul = FALSE,
  tlp = FALSE,
  i_value = 1:number,
  i_margin_left = "auto",
  i_margin_right = "auto",
  tlp_position = "top",
  tlp_color = "black",
  read_only = FALSE,
  hover = FALSE
)

Arguments

rating_id

The rating input id.

i_name

The icon name (fontawesome) or path (local).

i_lib

The icon library to use. Either "fontawesome" or "local".

number

The number of icons in the rating input.

init_value

The Initial value of the rating input.

i_on_color

The icon colors when selected (hex or Bootstrap 5 color).

i_off_color

The icon colors when deselected (hex or Bootstrap 5 color).

i_width

The icons width in css units.

i_height

The icons height in css units.

i_anim

The type of animations see Animate.css.

i_duration

The animation duration in seconds.

cumul

Logical, whether or not to use cumulative color for the icons.

tlp

Logical, whether or not to use a tooltip for the icon.

i_value

The character strings to be shown as message.

i_margin_left, i_margin_right

The length value for the margin that's either left or right of the icon. By default, "auto" is used for both properties. If space is needed on either side then a length of "0.2em" is recommended as a starting point.

tlp_position

The tooltip positions, "top", "right", "bottom", or "left".

tlp_color

The tooltip colors, "primary", "secondary", "light", "dark", "info", "warning", "danger", "success", "white", or "black".

read_only

Logical, whether or not to use a read only mode for the icons.

hover

Logical, whether or not to use a hover mode for the icons.

Value

A shiny.tag object for usage in ratingInput