Create a rating widget

Create a rating widget for usage in a Shiny UI.

ratingInput(
  inputId,
  label = NULL,
  width = "auto",
  number = 5,
  value = NA,
  cumul = FALSE,
  read_only = FALSE,
  hover = FALSE,
  i_on_color = "success",
  i_off_color = "dark",
  i_width = NULL,
  i_height = NULL,
  i_name = "star",
  i_lib = "fontawesome",
  i_anim = "none",
  i_duration = 2,
  i_value = 1:number,
  i_margin_left = "auto",
  i_margin_right = "auto",
  tlp = FALSE,
  tlp_position = "bottom",
  tlp_color = "black"
)

Arguments

inputId

The rating widget id.

label

Display label for the control.

width

The width of the input, e.g. '400px', or '100%'; see validateCssUnit()

number

The number of icons in the rating widget.

value

The Initial value.

cumul

Logical, whether or not to use cumulative color for the icons.

read_only

Logical, whether or not to use a read only mode for the icons.

hover

Logical, whether or not to use a hover mode for the icons.

i_on_color, i_off_color

The off and on icon colors (hex or Bootstrap 5 color).

i_width, i_height

The icons width and height in css units.

i_name

The icon name (fontawesome) or path (local).

i_lib

The icon library to use. Either "fontawesome" or "local".

i_anim

The type of animations see Animate.css.

i_duration

The animation duration in seconds.

i_value

The character strings to be shown as message.

i_margin_left, i_margin_right

The length value for the margin that's either left or right of the icon. By default, "auto" is used for both properties. If space is needed on either side then a length of "0.2em" is recommended as a starting point.

tlp

Logical, whether or not to use a tooltip for the icon.

tlp_position

The tooltip positions, "top", "right", "bottom", or "left".

tlp_color

The tooltip colors, "primary", "secondary", "light", "dark", "info", "warning", "danger", "success", "white", or "black".

Value

A rating widget for usage in Shiny UI.

Examples

ratingInput(
  inputId = "like",
  i_name = "heart",
  i_on_color = "danger"
)
#> <div class="form-group shiny-input-container" style="width:auto;">
#>   <label class="control-label m-0 p-0" id="like-label" for="like">
#>     <div></div>
#>     <div class="m-0 mt-2 p-0">
#>       <span id="like1" name="heart" lib="fontawesome" i_number="1" i_value="1" number="5" disabled="true" rating_id="like" i_on_color="var(--bs-danger)" i_off_color="var(--bs-dark)" cumulation="false" i_anim="none" type="button" class="m-0 p-0 action-button btn_rating " style="color: var(--bs-dark); --animate-duration: 2s; ;"><svg aria-hidden="true" role="img" viewBox="0 0 512 512" style="height:2em;width:2em;vertical-align:-0.125em;margin-left:auto;margin-right:auto;font-size:inherit;fill:currentColor;overflow:visible;position:relative;"><path d="M0 190.9V185.1C0 115.2 50.52 55.58 119.4 44.1C164.1 36.51 211.4 51.37 244 84.02L256 96L267.1 84.02C300.6 51.37 347 36.51 392.6 44.1C461.5 55.58 512 115.2 512 185.1V190.9C512 232.4 494.8 272.1 464.4 300.4L283.7 469.1C276.2 476.1 266.3 480 256 480C245.7 480 235.8 476.1 228.3 469.1L47.59 300.4C17.23 272.1 .0003 232.4 .0003 190.9L0 190.9z"/></svg></span>
#>       <span id="like2" name="heart" lib="fontawesome" i_number="2" i_value="2" number="5" disabled="true" rating_id="like" i_on_color="var(--bs-danger)" i_off_color="var(--bs-dark)" cumulation="false" i_anim="none" type="button" class="m-0 p-0 action-button btn_rating " style="color: var(--bs-dark); --animate-duration: 2s; ;"><svg aria-hidden="true" role="img" viewBox="0 0 512 512" style="height:2em;width:2em;vertical-align:-0.125em;margin-left:auto;margin-right:auto;font-size:inherit;fill:currentColor;overflow:visible;position:relative;"><path d="M0 190.9V185.1C0 115.2 50.52 55.58 119.4 44.1C164.1 36.51 211.4 51.37 244 84.02L256 96L267.1 84.02C300.6 51.37 347 36.51 392.6 44.1C461.5 55.58 512 115.2 512 185.1V190.9C512 232.4 494.8 272.1 464.4 300.4L283.7 469.1C276.2 476.1 266.3 480 256 480C245.7 480 235.8 476.1 228.3 469.1L47.59 300.4C17.23 272.1 .0003 232.4 .0003 190.9L0 190.9z"/></svg></span>
#>       <span id="like3" name="heart" lib="fontawesome" i_number="3" i_value="3" number="5" disabled="true" rating_id="like" i_on_color="var(--bs-danger)" i_off_color="var(--bs-dark)" cumulation="false" i_anim="none" type="button" class="m-0 p-0 action-button btn_rating " style="color: var(--bs-dark); --animate-duration: 2s; ;"><svg aria-hidden="true" role="img" viewBox="0 0 512 512" style="height:2em;width:2em;vertical-align:-0.125em;margin-left:auto;margin-right:auto;font-size:inherit;fill:currentColor;overflow:visible;position:relative;"><path d="M0 190.9V185.1C0 115.2 50.52 55.58 119.4 44.1C164.1 36.51 211.4 51.37 244 84.02L256 96L267.1 84.02C300.6 51.37 347 36.51 392.6 44.1C461.5 55.58 512 115.2 512 185.1V190.9C512 232.4 494.8 272.1 464.4 300.4L283.7 469.1C276.2 476.1 266.3 480 256 480C245.7 480 235.8 476.1 228.3 469.1L47.59 300.4C17.23 272.1 .0003 232.4 .0003 190.9L0 190.9z"/></svg></span>
#>       <span id="like4" name="heart" lib="fontawesome" i_number="4" i_value="4" number="5" disabled="true" rating_id="like" i_on_color="var(--bs-danger)" i_off_color="var(--bs-dark)" cumulation="false" i_anim="none" type="button" class="m-0 p-0 action-button btn_rating " style="color: var(--bs-dark); --animate-duration: 2s; ;"><svg aria-hidden="true" role="img" viewBox="0 0 512 512" style="height:2em;width:2em;vertical-align:-0.125em;margin-left:auto;margin-right:auto;font-size:inherit;fill:currentColor;overflow:visible;position:relative;"><path d="M0 190.9V185.1C0 115.2 50.52 55.58 119.4 44.1C164.1 36.51 211.4 51.37 244 84.02L256 96L267.1 84.02C300.6 51.37 347 36.51 392.6 44.1C461.5 55.58 512 115.2 512 185.1V190.9C512 232.4 494.8 272.1 464.4 300.4L283.7 469.1C276.2 476.1 266.3 480 256 480C245.7 480 235.8 476.1 228.3 469.1L47.59 300.4C17.23 272.1 .0003 232.4 .0003 190.9L0 190.9z"/></svg></span>
#>       <span id="like5" name="heart" lib="fontawesome" i_number="5" i_value="5" number="5" disabled="true" rating_id="like" i_on_color="var(--bs-danger)" i_off_color="var(--bs-dark)" cumulation="false" i_anim="none" type="button" class="m-0 p-0 action-button btn_rating " style="color: var(--bs-dark); --animate-duration: 2s; ;"><svg aria-hidden="true" role="img" viewBox="0 0 512 512" style="height:2em;width:2em;vertical-align:-0.125em;margin-left:auto;margin-right:auto;font-size:inherit;fill:currentColor;overflow:visible;position:relative;"><path d="M0 190.9V185.1C0 115.2 50.52 55.58 119.4 44.1C164.1 36.51 211.4 51.37 244 84.02L256 96L267.1 84.02C300.6 51.37 347 36.51 392.6 44.1C461.5 55.58 512 115.2 512 185.1V190.9C512 232.4 494.8 272.1 464.4 300.4L283.7 469.1C276.2 476.1 266.3 480 256 480C245.7 480 235.8 476.1 228.3 469.1L47.59 300.4C17.23 272.1 .0003 232.4 .0003 190.9L0 190.9z"/></svg></span>
#>     </div>
#>   </label>
#>   <input id="like" type="text" class="form-control rating_input d-none m-0 p-0" value number="5"/>
#> </div>