Create a rating widget

Create a rating widget for usage in a Shiny UI.

ratingInput(
  inputId,
  label = NULL,
  width = "auto",
  number = 5,
  value = NA,
  cumul = FALSE,
  read_only = FALSE,
  hover = FALSE,
  on_color = "success",
  off_color = "dark",
  size = 2,
  i_name = "star",
  i_lib = "font-awesome",
  anim = "none",
  duration = 2,
  tlp = FALSE,
  tlp_msg = 1:number,
  tlp_position = "bottom",
  tlp_color = "black"
)

Arguments

inputId

The rating widget id.

label

Display label for the control.

width

The width of the input, e.g. '400px', or '100%'; see validateCssUnit()

number

The number of icons in the rating widget.

value

The Initial value.

cumul

Logical, whether or not to use cumulative color for the icons.

read_only

Logical, whether or not to use a read only mode for the icons.

hover

Logical, whether or not to use a hover mode for the icons.

on_color

The icon colors when selected (hex or Bootstrap 5 color).

off_color

The icon colors when deselected (hex or Bootstrap 5 color).

size

The icon sizes (1 to 5).

i_name

The icon names.

i_lib

The icon library to use. Either "font-awesome" or "glyphicon".

anim

The type of animations see Animate.css.

duration

The animation duration in seconds.

tlp

Logical, whether or not to use a tooltip for the icon.

tlp_msg

The character strings to be shown as message in tooltips.

tlp_position

The tooltip positions, "top", "right", "bottom", or "left".

tlp_color

The tooltip colors, "primary", "secondary", "light", "dark", "info", "warning", "danger", "success", "white", or "black".

Value

A rating widget for usage in Shiny UI.

Examples

ratingInput(
inputId = "like",
i_name = "heart",
on_color = "danger"
)
#> <div class="form-group shiny-input-container" style="width:auto;">
#>   <label class="control-label m-0 p-0" id="like-label" for="like">
#>     <div></div>
#>     <div class="m-0 mt-2 p-0">
#>       <i class="fa fa-heart m-0 p-0 action-button btn_rating fa-2x " role="presentation" aria-label="heart icon" id="like1" i_number="1" number="5" disabled="true" rating_id="like" on_color="var(--bs-danger)" off_color="var(--bs-dark)" cumulation="false" anim="none" type="button" style="color: var(--bs-dark); --animate-duration: 2s; "></i>
#>       <i class="fa fa-heart m-0 p-0 action-button btn_rating fa-2x " role="presentation" aria-label="heart icon" id="like2" i_number="2" number="5" disabled="true" rating_id="like" on_color="var(--bs-danger)" off_color="var(--bs-dark)" cumulation="false" anim="none" type="button" style="color: var(--bs-dark); --animate-duration: 2s; "></i>
#>       <i class="fa fa-heart m-0 p-0 action-button btn_rating fa-2x " role="presentation" aria-label="heart icon" id="like3" i_number="3" number="5" disabled="true" rating_id="like" on_color="var(--bs-danger)" off_color="var(--bs-dark)" cumulation="false" anim="none" type="button" style="color: var(--bs-dark); --animate-duration: 2s; "></i>
#>       <i class="fa fa-heart m-0 p-0 action-button btn_rating fa-2x " role="presentation" aria-label="heart icon" id="like4" i_number="4" number="5" disabled="true" rating_id="like" on_color="var(--bs-danger)" off_color="var(--bs-dark)" cumulation="false" anim="none" type="button" style="color: var(--bs-dark); --animate-duration: 2s; "></i>
#>       <i class="fa fa-heart m-0 p-0 action-button btn_rating fa-2x " role="presentation" aria-label="heart icon" id="like5" i_number="5" number="5" disabled="true" rating_id="like" on_color="var(--bs-danger)" off_color="var(--bs-dark)" cumulation="false" anim="none" type="button" style="color: var(--bs-dark); --animate-duration: 2s; "></i>
#>     </div>
#>   </label>
#>   <input id="like" type="number" class="form-control rating_input d-none m-0 p-0" value="NA" min="1" max="5" number="5"/>
#> </div>